+++ jax/_src/numpy/lax_numpy.py
@@ -60,7 +60,7 @@ from jax._src.numpy.sorting import argsort, sort
 from jax._src.numpy.vectorize import vectorize
 from jax._src.sharding_impls import canonicalize_sharding
 from jax._src.typing import (
-  Array, ArrayLike, DType, DTypeLike, DeprecatedArg, DimSize, Shape, SupportsShape
+  Array, ArrayLike, DType, DTypeLike, DimSize, Shape, SupportsShape
 )
 from jax._src.util import (
     canonicalize_axis as _canonicalize_axis,
@@ -3343,10 +3343,6 @@ def clip(
   /,
   min: ArrayLike | None = None,
   max: ArrayLike | None = None,
-  *,
-  a: ArrayLike | DeprecatedArg = DeprecatedArg(),
-  a_min: ArrayLike | None | DeprecatedArg = DeprecatedArg(),
-  a_max: ArrayLike | None | DeprecatedArg = DeprecatedArg()
 ) -> Array:
   """Clip array values to a specified range.
 
@@ -3360,12 +3356,6 @@ def clip(
     max: optional maximum value of the clipped range; if ``None`` (default) then
       result will not be clipped to any maximum value. If specified, it should be
       broadcast-compatible with ``arr`` and ``min``.
-    a: deprecated alias of the ``arr`` argument.  Will result in a
-      :class:`DeprecationWarning` if used.
-    a_min: deprecated alias of the ``min`` argument. Will result in a
-      :class:`DeprecationWarning` if used.
-    a_max: deprecated alias of the ``max`` argument. Will result in a
-      :class:`DeprecationWarning` if used.
 
   Returns:
     An array containing values from ``arr``, with values smaller than ``min`` set
@@ -3381,19 +3371,8 @@ def clip(
     >>> jnp.clip(arr, 2, 5)
     Array([2, 2, 2, 3, 4, 5, 5, 5], dtype=int32)
   """
-  # TODO(micky774): deprecated 2024-4-2, remove after deprecation expires.
-  arr = a if not isinstance(a, DeprecatedArg) else arr
   if arr is None:
     raise ValueError("No input was provided to the clip function.")
-  min = a_min if not isinstance(a_min, DeprecatedArg) else min
-  max = a_max if not isinstance(a_max, DeprecatedArg) else max
-  if any(not isinstance(t, DeprecatedArg) for t in (a, a_min, a_max)):
-    deprecations.warn(
-      "jax-numpy-clip-args",
-      ("Passing arguments 'a', 'a_min' or 'a_max' to jax.numpy.clip is "
-       "deprecated. Please use 'arr', 'min' or 'max' respectively instead."),
-      stacklevel=2,
-    )
 
   util.check_arraylike("clip", arr)
   if any(iscomplexobj(t) for t in (arr, min, max)):
@@ -4694,8 +4673,7 @@ def vstack(tup: np.ndarray | Array | Sequence[ArrayLike],
   if isinstance(tup, (np.ndarray, Array)):
     arrs = api.vmap(atleast_2d)(tup)
   else:
-    # TODO(jakevdp): Non-array input deprecated 2023-09-22; change to error.
-    util.check_arraylike("vstack", *tup, emit_warning=True)
+    util.check_arraylike("vstack", *tup)
     arrs = [atleast_2d(m) for m in tup]
   return concatenate(arrs, axis=0, dtype=dtype)
 
@@ -4754,8 +4732,7 @@ def hstack(tup: np.ndarray | Array | Sequence[ArrayLike],
     arrs = api.vmap(atleast_1d)(tup)
     arr0_ndim = arrs.ndim - 1
   else:
-    # TODO(jakevdp): Non-array input deprecated 2023-09-22; change to error.
-    util.check_arraylike("hstack", *tup, emit_warning=True)
+    util.check_arraylike("hstack", *tup)
     arrs = [atleast_1d(m) for m in tup]
     arr0_ndim = arrs[0].ndim
   return concatenate(arrs, axis=0 if arr0_ndim == 1 else 1, dtype=dtype)
@@ -4816,8 +4793,7 @@ def dstack(tup: np.ndarray | Array | Sequence[ArrayLike],
   if isinstance(tup, (np.ndarray, Array)):
     arrs = api.vmap(atleast_3d)(tup)
   else:
-    # TODO(jakevdp): Non-array input deprecated 2023-09-22; change to error.
-    util.check_arraylike("dstack", *tup, emit_warning=True)
+    util.check_arraylike("dstack", *tup)
     tup = util.ensure_arraylike_tuple("dstack", tup)
     arrs = [atleast_3d(m) for m in tup]
   return concatenate(arrs, axis=2, dtype=dtype)
@@ -4878,8 +4854,7 @@ def column_stack(tup: np.ndarray | Array | Sequence[ArrayLike]) -> Array:
   if isinstance(tup, (np.ndarray, Array)):
     arrs = api.vmap(lambda x: atleast_2d(x).T)(tup) if tup.ndim < 3 else tup
   else:
-    # TODO(jakevdp): Non-array input deprecated 2023-09-22; change to error.
-    util.check_arraylike("column_stack", *tup, emit_warning=True)
+    util.check_arraylike("column_stack", *tup)
     arrs = [atleast_2d(arr).T if arr.ndim < 2 else arr for arr in map(asarray, tup)]
   return concatenate(arrs, axis=1)
 
@@ -5132,7 +5107,7 @@ def atleast_1d(*arys: ArrayLike) -> Array | list[Array]:
     >>> jnp.atleast_1d(x, y)
     [Array([1.], dtype=float32), Array([0, 1, 2, 3], dtype=int32)]
   """
-  util.check_arraylike("atleast_1d", *arys, emit_warning=True)
+  util.check_arraylike("atleast_1d", *arys)
   if len(arys) == 1:
     return array(arys[0], copy=False, ndmin=1)
   else:
@@ -5195,8 +5170,7 @@ def atleast_2d(*arys: ArrayLike) -> Array | list[Array]:
     >>> jnp.atleast_2d(x, y)
     [Array([[1.]], dtype=float32), Array([[0, 1, 2, 3]], dtype=int32)]
   """
-  # TODO(jakevdp): Non-array input deprecated 2023-09-22; change to error.
-  util.check_arraylike("atleast_2d", *arys, emit_warning=True)
+  util.check_arraylike("atleast_2d", *arys)
   if len(arys) == 1:
     return array(arys[0], copy=False, ndmin=2)
   else:
@@ -5265,8 +5239,7 @@ def atleast_3d(*arys: ArrayLike) -> Array | list[Array]:
       [2]
       [3]]]
   """
-  # TODO(jakevdp): Non-array input deprecated 2023-09-22; change to error.
-  util.check_arraylike("atleast_3d", *arys, emit_warning=True)
+  util.check_arraylike("atleast_3d", *arys)
   if len(arys) == 1:
     arr = asarray(arys[0])
     if arr.ndim == 0: